Modify validation for upcoming PR
